move more methods over to Diagnostic
