+++ crates/ruff_linter/src/linter.rs
@@ -516,7 +516,9 @@ pub fn lint_only(
 
     LinterResult {
         has_valid_syntax: parsed.has_valid_syntax(),
-        has_no_syntax_errors: !diagnostics.iter().any(OldDiagnostic::is_syntax_error),
+        has_no_syntax_errors: !diagnostics
+            .iter()
+            .any(|diagnostic| diagnostic.is_syntax_error()),
         diagnostics,
     }
 }
@@ -630,7 +632,9 @@ pub fn lint_fix<'a>(
 
         if iterations == 0 {
             has_valid_syntax = parsed.has_valid_syntax();
-            has_no_syntax_errors = !diagnostics.iter().any(OldDiagnostic::is_syntax_error);
+            has_no_syntax_errors = !diagnostics
+                .iter()
+                .any(|diagnostic| diagnostic.is_syntax_error());
         } else {
             // If the source code had no syntax errors on the first pass, but
             // does on a subsequent pass, then we've introduced a
